fix(storage): keep chunk metadata parsing from accepting short reads
