+++ src/EventStore.Core/TransactionLog/Chunks/TFChunk/ChunkLocalFileSystem.cs
@@ -5,7 +5,6 @@ using System.Threading.Tasks;
 using DotNext.Buffers;
 using EventStore.Core.Exceptions;
 using EventStore.Core.TransactionLog.FileNamingStrategy;
-using Microsoft.Win32.SafeHandles;
 
 namespace EventStore.Core.TransactionLog.Chunks.TFChunk;
 
@@ -42,34 +41,21 @@ public sealed class ChunkLocalFileSystem(IVersionedFileNamingStrategy namingStra
 
 	public async ValueTask<ChunkHeader> ReadHeaderAsync(string fileName, CancellationToken token)
 	{
-		using var handle = OpenValidatedReadHandle(fileName, out _);
+		using var handle = ChunkFileReadHelper.OpenValidatedMetadataReadHandle(fileName, out _);
 
 		using var buffer = Memory.AllocateExactly<byte>(ChunkHeader.Size);
-		await RandomAccess.ReadAsync(handle, buffer.Memory, 0L, token);
+		await ChunkFileReadHelper.ReadExactlyAsync(handle, buffer.Memory, 0L, fileName, token);
 		return new(buffer.Span);
 	}
 
 	public async ValueTask<ChunkFooter> ReadFooterAsync(string fileName, CancellationToken token)
 	{
-		using var handle = OpenValidatedReadHandle(fileName, out var length);
+		using var handle = ChunkFileReadHelper.OpenValidatedMetadataReadHandle(fileName, out var length);
 
 		using var buffer = Memory.AllocateExactly<byte>(ChunkFooter.Size);
-		await RandomAccess.ReadAsync(handle, buffer.Memory, length - ChunkFooter.Size, token);
+		await ChunkFileReadHelper.ReadExactlyAsync(handle, buffer.Memory, length - ChunkFooter.Size, fileName, token);
 		return new(buffer.Span);
 	}
 
 	public IChunkEnumerator CreateChunkEnumerator() => new TFChunkEnumerator(NamingStrategy);
-
-	private static SafeFileHandle OpenValidatedReadHandle(string fileName, out long length)
-	{
-		var handle = File.OpenHandle(fileName, FileMode.Open, FileAccess.Read, FileShare.ReadWrite,
-			FileOptions.Asynchronous);
-		length = RandomAccess.GetLength(handle);
-		if (length >= ChunkFooter.Size + ChunkHeader.Size)
-			return handle;
-
-		handle.Dispose();
-		throw new CorruptDatabaseException(new BadChunkInDatabaseException(
-			$"Chunk file '{fileName}' is bad. It does not have enough size for header and footer. File size is {length} bytes."));
-	}
 }
